+++ src/main/java/org/com/deputatbot/repos/OkrugNduRepo.java
@@ -7,7 +7,7 @@ import org.com.deputatbot.domain.OkrugNdu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface OkrugNduRepo extends JpaRepository<OkrugNdu, Integer> {
+public interface OkrugNduRepo extends JpaRepository<OkrugNdu, Long> {
     OkrugNdu findByNumber(Integer number);
     OkrugNdu findByDeputat(Deputat deputat);
 }
